# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meridian Advisory API",
    "description": "Consultation booking, payments, client and consultant portals, commerce and content APIs.\n\nAll responses use a single envelope: `{ success: true, data }` or `{ success: false, error: { code, message } }`.\n\nAuthenticate with `Authorization: Bearer <accessToken>`. Access tokens are short-lived; refresh them at `POST /api/auth/refresh`, which reads an httpOnly cookie.",
    "version": "1.0.0",
    "contact": {
      "name": "Support",
      "email": "support@meridianadvisory.co.ke"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/services/{slug}": {
      "get": {
        "summary": "Service detail with consultants",
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/consultants/{slug}": {
      "get": {
        "summary": "Consultant profile with services and reviews",
        "tags": [
          "Consultants"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "development"
    }
  ],
...
}